@@ -23,7 +23,6 @@ package maxcluster
 import (
 	"encoding/gob"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"testing"
@@ -87,7 +86,7 @@ func init() {
 		panic(err)
 	}
 
-	token, err = ioutil.ReadFile(config.ManagerTokenFile)
+	token, err = os.ReadFile(config.ManagerTokenFile)
 	if err != nil {
 		panic(err)
 	}
