@@ -21,6 +21,7 @@ package jobqueue
 // This file contains the web interface code of the server.
 
 import (
+	"embed"
 	"net/http"
 	"strings"
 
@@ -30,6 +31,9 @@ import (
 	"github.com/VertebrateResequencing/wr/queue"
 	"github.com/gorilla/websocket"
 )
+
+//go:embed static
+var staticFS embed.FS
 
 // jstatusReq is what the status webpage sends us to ask for info about jobs.
 type jstatusReq struct {
@@ -103,40 +107,38 @@ type JStatus struct {
 	Exited        bool
 }
 
-// webInterfaceStatic is a http handler for our static documents in static.go
-// (which in turn come from the static folder in the git repository). static.go
-// is auto-generated by:
-// $ esc -pkg jobqueue -prefix static -private -o jobqueue/static.go static
+// webInterfaceStatic is a http handler for our static documents in the static
+// folder of the source code repository, which are embedded at compile time.
 func webInterfaceStatic(s *Server) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		// our home page is /status.html
 		path := r.URL.Path
-		if path == "/" || path == "/status" {
-			path = "/status.html"
+		path = strings.TrimPrefix(path, "/")
+		if path == "" || path == "status" {
+			path = "status.html"
+		}
 
+		if path == "status.html" {
 			ok := s.httpAuthorized(w, r)
 			if !ok {
 				return
 			}
 		}
 
-		// during development, to avoid having to rebuild and restart manager on
-		// every change to a file in static dir, do:
-		// $ esc -pkg jobqueue -prefix $PWD/static -private -o jobqueue/static.go $PWD/static
-		// and set the boolean to true. Don't forget to rerun esc without the abs
-		// paths and change the boolean back to false before any commit!
-		doc, err := _escFSByte(false, path)
+		path = "static/" + path
+		doc, err := staticFS.ReadFile(path)
 		if err != nil {
+			s.Warn("not found", "err", err)
 			http.NotFound(w, r)
 			return
 		}
 
 		switch {
-		case strings.HasPrefix(path, "/js"):
+		case strings.HasPrefix(path, "static/js"):
 			w.Header().Set("Content-Type", "application/json; charset=utf-8")
-		case strings.HasPrefix(path, "/css"):
+		case strings.HasPrefix(path, "static/css"):
 			w.Header().Set("Content-Type", "text/css; charset=utf-8")
-		case strings.HasPrefix(path, "/fonts"):
+		case strings.HasPrefix(path, "static/fonts"):
 			switch {
 			case strings.HasSuffix(path, ".eot"):
 				w.Header().Set("Content-Type", "application/vnd.ms-fontobject")
